[Feat]:BidRepository 업데이트

+++ src/main/java/com/backend/domain/bid/dto/BidCurrentResponseDto.java
@@ -8,7 +8,7 @@ public record BidCurrentResponseDto (
         String productName,
         Long currentPrice,
         Long initialPrice,
-        Long bidCount,
+        Integer bidCount,
         String status,
         LocalDateTime endTime,
         String remainingTime,
